requirements.txt
autopep8==1.6.0
bcrypt==3.2.2
bidict==0.22.0
cachelib==0.9.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
deta==1.1.0
Flask==2.1.3
Flask-Session==0.4.0
Flask-SocketIO==5.2.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pycodestyle==2.8.0
pycparser==2.21
python-decouple==3.6
python-engineio==4.3.3
python-socketio==5.7.0
toml==0.10.2
Werkzeug==2.1.2
zipp==3.8.1
